+++ examples/vite-simple/main.js
@@ -1,10 +1,21 @@
 import './style.css'
 // import 3 different sizes of the image and create a srcset from them
-import srcsetAvif from '../example.jpg?w=500;700;900;1200&avif&srcset'
+import srcsetAvif from '../example.jpg?w=500;700;900;1200&format=avif&as=srcset'
 // do it a second time, but now as webp since safari can't display avif
-import srcsetWebp from '../example.jpg?w=500;700;900;1200&webp&srcset'
+import srcsetWebp from '../example.jpg?w=500;700;900;1200&format=webp&as=srcset'
 // create a small placeholder and import its metadata
-import { src as placeholder, width, height } from '../example.jpg?width=300&metadata'
+import { src as placeholder, width, height } from '../example.jpg?w=300&as=metadata'
+
+const sponsors = import.meta.glob('./sponsors/*', {
+  query: { format: 'webp', w: 100 }
+})
+
+let sponsorHtml = ''
+for (let sponsor of Object.values(sponsors)) {
+  const import_statment = sponsor()
+  const url = (await import_statment).default
+  sponsorHtml += `<img src="${url}" /><br/>`
+}
 
 document.querySelector('#app').innerHTML = `
   <h1>Hello Imagetools!</h1>
@@ -20,4 +31,10 @@ document.querySelector('#app').innerHTML = `
         height="${height}"
         alt="Women Lying Near to a Multicolored Glass Window Close-up Photography"/>
   </picture>
+
+  <div style="width:300px; margin: 50px auto">
+    Here are a few sponsors of Vite:<br/>
+    <br/>
+    ${sponsorHtml}
+  </div>
 `