+++ packages/vite/src/__tests__/main.test.ts
@@ -309,7 +309,7 @@ describe('vite-imagetools', () => {
           build: { write: false },
           plugins: [
             testEntry(`
-                            import Image from "./with-metadata.png?metadata"
+                            import Image from "./with-metadata.png?as=metadata"
                             window.__IMAGE__ = Image
                         `),
             imagetools({
@@ -332,7 +332,7 @@ describe('vite-imagetools', () => {
           build: { write: false },
           plugins: [
             testEntry(`
-                            import Image from "./with-metadata.png?metadata"
+                            import Image from "./with-metadata.png?as=metadata"
                             window.__IMAGE__ = Image
                         `),
             imagetools({
@@ -357,12 +357,12 @@ describe('vite-imagetools', () => {
           build: { write: false },
           plugins: [
             testEntry(`
-                            import Image from "./with-metadata.png?metadata"
+                            import Image from "./with-metadata.png?as=metadata"
                             window.__IMAGE__ = Image
                         `),
             imagetools({
               resolveConfigs() {
-                return [{ width: '300' }, { width: '500' }]
+                return [{ w: '300' }, { w: '500' }]
               }
             })
           ]
@@ -385,7 +385,7 @@ describe('vite-imagetools', () => {
                             window.__IMAGE__ = Image
                         `),
             imagetools({
-              defaultDirectives: new URLSearchParams('width=300;500')
+              defaultDirectives: new URLSearchParams('w=300;500')
             })
           ]
         })) as RollupOutput | RollupOutput[]
@@ -407,7 +407,7 @@ describe('vite-imagetools', () => {
             imagetools({
               defaultDirectives: (id) => {
                 if (id.searchParams.has('mypreset')) {
-                  return new URLSearchParams('width=300;500')
+                  return new URLSearchParams('w=300;500')
                 }
                 return new URLSearchParams()
               }
@@ -432,7 +432,7 @@ describe('vite-imagetools', () => {
             imagetools({
               defaultDirectives: (id) => {
                 if (id.searchParams.has('mypreset')) {
-                  return new URLSearchParams('metadata')
+                  return new URLSearchParams('as=metadata')
                 }
                 return new URLSearchParams()
               }
@@ -556,7 +556,7 @@ describe('vite-imagetools', () => {
       build: { write: false },
       plugins: [
         testEntry(`
-                    import Image from "./pexels-allec-gomes-5195763.png?metadata"
+                    import Image from "./pexels-allec-gomes-5195763.png?as=metadata"
                     window.__IMAGE__ = Image
                 `),
         imagetools()
@@ -587,7 +587,7 @@ describe('vite-imagetools', () => {
       build: { write: false },
       plugins: [
         testEntry(`
-                    import { width, height, format } from "./pexels-allec-gomes-5195763.png?metadata"
+                    import { width, height, format } from "./pexels-allec-gomes-5195763.png?as=metadata"
                     window.__IMAGE__ = { width, height, format }
                 `),
         imagetools()
@@ -610,7 +610,7 @@ describe('vite-imagetools', () => {
       build: { write: false },
       plugins: [
         testEntry(`
-                    import { width, format } from "./pexels-allec-gomes-5195763.png?metadata=width;format"
+                    import { width, format } from "./pexels-allec-gomes-5195763.png?as=metadata:width;format"
                     window.__IMAGE__ = { width, format }
                 `),
         imagetools()
@@ -633,7 +633,7 @@ describe('vite-imagetools', () => {
       build: { write: false },
       plugins: [
         testEntry(`
-                        import Image from "./with-metadata.png?srcset"
+                        import Image from "./with-metadata.png?as=srcset"
                         window.__IMAGE__ = Image
                     `),
         imagetools()
@@ -645,6 +645,32 @@ describe('vite-imagetools', () => {
     window.eval(files[0].code)
 
     expect(window.__IMAGE__).toBe('/assets/with-metadata-404f605d.png 600w')
+  })
+
+  test('async output format', async () => {
+    const bundle = (await build({
+      root: join(__dirname, '__fixtures__'),
+      logLevel: 'warn',
+      build: { write: false },
+      plugins: [
+        testEntry(`
+          import Image from "./with-metadata.png?as=run"
+          window.__IMAGE__ = Image
+        `),
+        imagetools({
+          extendOutputFormats: (defaults) => ({
+            ...defaults,
+            run: () => () => new Promise((resolve) => setTimeout(() => resolve('success'), 500))
+          })
+        })
+      ]
+    })) as RollupOutput | RollupOutput[]
+
+    const files = getFiles(bundle, '**.js') as OutputChunk[]
+    const { window } = new JSDOM(``, { runScripts: 'outside-only' })
+    window.eval(files[0].code)
+
+    expect(window.__IMAGE__).toBe('success')
   })
 
   describe('utils', () => {